refactor(db): migrate persistence to postgres

+++ src/lib/server/db/migrations/003_lucia_auth.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS sessions (
     id TEXT PRIMARY KEY,
     user_id INTEGER NOT NULL,
-    expires_at INTEGER NOT NULL,
+    expires_at BIGINT NOT NULL,
     FOREIGN KEY (user_id) REFERENCES users(id) ON DELETE CASCADE
 );
 CREATE INDEX IF NOT EXISTS idx_sessions_user_id ON sessions(user_id);
